+++ book/src/SUMMARY.md
@@ -8,6 +8,7 @@
   - [Writing html!](./html-macro/README.md)
     - [Writing html!](./html-macro/html-macro.md)
     - [Compile Time Errors](./html-macro/compile-time-errors.md)
+    - [Working with Text](./html-macro/text/README.md)
   - [Virtual DOM](./virtual-dom/README.md)
     - [Unit Testing your Views](./virtual-dom/unit-testing-views.md)
   - [Server Side Rendering (SSR)](./views/server-side-rendering/README.md)
