+++ book/src/html-macro/text/README.md
@@ -1,0 +1,10 @@
+# Working with Text
+
+Rather than needing you to wrap your text in `"` quotation marks,
+the `html-macro` will work with raw unquoted text.
+
+Here are some examples of how the `html!` macro works with text:
+
+```rust
+{{ #include ../../../../crates/html-macro-test/src/text.rs }}
+```
