+++ crates/html-macro-test/src/lib.rs
@@ -1,9 +1,10 @@
 #![feature(proc_macro_hygiene)]
-#![cfg(test)]
 
 use html_macro::html;
 use std::collections::HashMap;
 use virtual_node::{IterableNodes, VElement, VirtualNode};
+
+mod text;
 
 struct HtmlMacroTest<'a> {
     desc: &'a str,
@@ -209,135 +210,10 @@ fn vec_of_nodes() {
     .test();
 }
 
-// FIXME: Move text tests into a separate file and import into book under
-// html! section
-
-#[test]
-fn text_root_node() {
-    HtmlMacroTest {
-        desc: "Text as root node",
-        generated: html! { some text },
-        expected: VirtualNode::text("some text"),
-    }
-    .test()
-}
-
 /// Just make sure that this compiles since type is a keyword
 #[test]
 fn type_attribute() {
     html! { <link rel="stylesheet" type="text/css" href="/app.css" /> };
-}
-
-#[test]
-fn text_variable_root() {
-    let text = "hello world";
-
-    HtmlMacroTest {
-        desc: "Text variable root",
-        generated: html! { { text } },
-        expected: VirtualNode::text("hello world"),
-    }
-    .test()
-}
-
-#[test]
-fn text_variable_child() {
-    let text = "world";
-
-    assert_eq!(
-        &html! { <div>{ text }</div> }.to_string(),
-        "<div>world</div>"
-    )
-}
-
-#[test]
-fn text_space_after_start_tag() {
-    assert_eq!(
-        &html! { <div> After Start Tag</div> }.to_string(),
-        "<div> After Start Tag</div>"
-    )
-}
-
-#[test]
-fn text_space_before_end_tag() {
-    assert_eq!(
-        &html! { <div>Before End Tag </div> }.to_string(),
-        "<div>Before End Tag </div>"
-    )
-}
-
-#[test]
-fn text_space_before_block() {
-    let text = "Before Block";
-
-    assert_eq!(
-        &html! { <div> {text}</div> }.to_string(),
-        "<div> Before Block</div>"
-    )
-}
-
-#[test]
-fn text_space_after_block() {
-    let text = "Hello";
-
-    assert_eq!(
-        &html! { <div>{text} </div> }.to_string(),
-        "<div>Hello </div>"
-    )
-}
-
-#[test]
-fn text_space_in_block_ignored() {
-    let text = "Hello";
-
-    assert_eq!(
-        &html! { <div>{ text }</div> }.to_string(),
-        "<div>Hello</div>"
-    )
-}
-
-#[test]
-fn text_multiple_text_no_space_between() {
-    let hello = "Hello";
-    let world = "World";
-
-    assert_eq!(
-        &html! { <div>{ hello }{ world }</div> }.to_string(),
-        "<div>HelloWorld</div>"
-    )
-}
-
-#[test]
-fn text_multiple_text_space_between() {
-    let hello = "Hello";
-    let world = "World";
-
-    assert_eq!(
-        &html! { <div>{ hello } { world }</div> }.to_string(),
-        "<div>Hello World</div>"
-    )
-}
-
-#[test]
-fn text_multiple_text_space_around() {
-    let hello = "Hello";
-    let world = "World";
-
-    assert_eq!(
-        &html! { <div> { hello }{ world } </div> }.to_string(),
-        "<div> HelloWorld </div>"
-    )
-}
-
-#[test]
-fn text_multiple_text_space_between_around() {
-    let hello = "Hello";
-    let world = "World";
-
-    assert_eq!(
-        &html! { <div> { hello } { world } </div> }.to_string(),
-        "<div> Hello World </div>"
-    )
 }
 
 // Verify that all of our self closing tags work as both.
