+++ crates/html-macro/src/parser/mod.rs
@@ -67,7 +67,7 @@ impl HtmlParser {
                 attrs,
                 closing_bracket_span,
             } => {
-                self.parse_open_tag(name, attrs);
+                self.parse_open_tag(name, closing_bracket_span, attrs);
                 self.last_tag_kind = Some(TagKind::Open);
             }
             Tag::Close {
